split model files into smaller files
