+++ front/lib/api/extract.ts
@@ -1,8 +1,9 @@
 import { Op } from "sequelize";
 
 import { Authenticator } from "@app/lib/auth";
+import { ModelId } from "@app/lib/databases";
 import { isDevelopmentOrDustWorkspace } from "@app/lib/development";
-import { EventSchema, ExtractedEvent, ModelId } from "@app/lib/models";
+import { EventSchema, ExtractedEvent } from "@app/lib/models";
 import { generateModelSId } from "@app/lib/utils";
 import { EventSchemaType, ExtractedEventType } from "@app/types/extract";
