+++ front/lib/models/apps.ts
@@ -1,0 +1,289 @@
+import {
+  CreationOptional,
+  DataTypes,
+  ForeignKey,
+  InferAttributes,
+  InferCreationAttributes,
+  Model,
+} from "sequelize";
+
+import { front_sequelize } from "@app/lib/databases";
+import { Workspace } from "@app/lib/models/workspace";
+
+export class App extends Model<
+  InferAttributes<App>,
+  InferCreationAttributes<App>
+> {
+  declare id: CreationOptional<number>;
+  declare createdAt: CreationOptional<Date>;
+  declare updatedAt: CreationOptional<Date>;
+
+  declare sId: string;
+  declare name: string;
+  declare description?: string;
+  declare visibility: "public" | "private" | "unlisted" | "deleted";
+  declare savedSpecification?: string;
+  declare savedConfig?: string;
+  declare savedRun?: string;
+  declare dustAPIProjectId: string;
+
+  declare workspaceId: ForeignKey<Workspace["id"]>;
+}
+App.init(
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    sId: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    description: {
+      type: DataTypes.STRING,
+    },
+    visibility: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    savedSpecification: {
+      type: DataTypes.TEXT,
+    },
+    savedConfig: {
+      type: DataTypes.TEXT,
+    },
+    savedRun: {
+      type: DataTypes.TEXT,
+    },
+    dustAPIProjectId: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+  },
+  {
+    modelName: "app",
+    sequelize: front_sequelize,
+    indexes: [
+      { unique: true, fields: ["sId"] },
+      { fields: ["workspaceId", "visibility"] },
+      { fields: ["workspaceId", "sId", "visibility"] },
+    ],
+  }
+);
+Workspace.hasMany(App);
+
+export class Provider extends Model<
+  InferAttributes<Provider>,
+  InferCreationAttributes<Provider>
+> {
+  declare id: CreationOptional<number>;
+  declare createdAt: CreationOptional<Date>;
+  declare updatedAt: CreationOptional<Date>;
+
+  declare providerId: string;
+  declare config: string;
+
+  declare workspaceId: ForeignKey<Workspace["id"]>;
+}
+Provider.init(
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    providerId: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    config: {
+      type: DataTypes.TEXT,
+      allowNull: false,
+    },
+  },
+  {
+    modelName: "provider",
+    sequelize: front_sequelize,
+    indexes: [{ fields: ["workspaceId"] }],
+  }
+);
+Workspace.hasMany(Provider);
+
+export class Dataset extends Model<
+  InferAttributes<Dataset>,
+  InferCreationAttributes<Dataset>
+> {
+  declare id: CreationOptional<number>;
+  declare createdAt: CreationOptional<Date>;
+  declare updatedAt: CreationOptional<Date>;
+
+  declare name: string;
+  declare description?: string;
+
+  declare workspaceId: ForeignKey<Workspace["id"]>;
+  declare appId: ForeignKey<App["id"]>;
+}
+Dataset.init(
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    description: {
+      type: DataTypes.STRING,
+    },
+  },
+  {
+    modelName: "dataset",
+    sequelize: front_sequelize,
+    indexes: [{ fields: ["workspaceId", "appId", "name"] }],
+  }
+);
+App.hasMany(Dataset);
+Workspace.hasMany(Dataset);
+
+export class Clone extends Model<
+  InferAttributes<Clone>,
+  InferCreationAttributes<Clone>
+> {
+  declare id: CreationOptional<number>;
+  declare createdAt: CreationOptional<Date>;
+  declare updatedAt: CreationOptional<Date>;
+
+  declare fromId: ForeignKey<App["id"]>;
+  declare toId: ForeignKey<App["id"]>;
+}
+Clone.init(
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      primaryKey: true,
+      autoIncrement: true,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    fromId: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      references: {
+        model: "apps",
+        key: "id",
+      },
+    },
+    toId: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      references: {
+        model: "apps",
+        key: "id",
+      },
+    },
+  },
+  {
+    modelName: "clone",
+    sequelize: front_sequelize,
+  }
+);
+Clone.belongsTo(App, { foreignKey: "fromId" });
+Clone.belongsTo(App, { foreignKey: "toId" });
+
+export class Run extends Model<
+  InferAttributes<Run>,
+  InferCreationAttributes<Run>
+> {
+  declare id: CreationOptional<number>;
+  declare createdAt: CreationOptional<Date>;
+  declare updatedAt: CreationOptional<Date>;
+
+  declare dustRunId: string;
+  declare runType: string;
+
+  declare appId: ForeignKey<App["id"]>;
+  declare workspaceId: ForeignKey<Workspace["id"]>;
+}
+
+Run.init(
+  {
+    id: {
+      type: DataTypes.INTEGER,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+      allowNull: false,
+      defaultValue: DataTypes.NOW,
+    },
+    dustRunId: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    runType: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+  },
+  {
+    modelName: "run",
+    sequelize: front_sequelize,
+    indexes: [
+      { fields: ["workspaceId", "appId", "runType", "createdAt"] },
+      { unique: true, fields: ["dustRunId"] },
+    ],
+  }
+);
+App.hasMany(Run);
+Workspace.hasMany(Run);
