+++ front/types/extract.ts
@@ -1,4 +1,4 @@
-import { ModelId } from "@app/lib/models";
+import { ModelId } from "@app/lib/databases";
 
 export type EventSchemaStatus = "active" | "disabled";
 export type ExtractedEventStatus = "pending" | "accepted" | "rejected";
