+++ front/types/membership_invitation.ts
@@ -1,4 +1,4 @@
-import { ModelId } from "@app/lib/models";
+import { ModelId } from "@app/lib/databases";
 
 export type MembershipInvitationType = {
   id: ModelId;
